+++ frontend/src/AdminPanel/pages/AdminPublications.jsx
@@ -3,6 +3,7 @@ import axios from "axios";
 import { FaEdit, FaTrash, FaBookOpen } from "react-icons/fa"; // Importing icons
 import "../style/adminpublications.css";
 import { toast } from "react-toastify";
+import Swal from "sweetalert2";
 
 export default function AdminPublications() {
   const [publications, setPublications] = useState([]);
@@ -22,7 +23,7 @@ export default function AdminPublications() {
       await axios.put(`http://localhost:5000/api/publications/${editId}`, formData);
       setEditId(null);
       toast.success("Data updated successfully!")
-      
+
     } else {
       await axios.post("http://localhost:5000/api/publications", formData);
     }
@@ -32,9 +33,21 @@ export default function AdminPublications() {
   };
 
   const handleDelete = async (id) => {
-    await axios.delete(`http://localhost:5000/api/publications/${id}`);
-    setPublications(publications.filter((item) => item._id !== id));
-    toast.delete("Data deleted successfully!")
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#d33",
+      cancelButtonColor: "#3085d6",
+      confirmButtonText: "Yes, delete it!",
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        await axios.delete(`http://localhost:5000/api/publications/${id}`);
+        setPublications(publications.filter((item) => item._id !== id));
+        toast.success("Data deleted successfully!")
+      }
+    })
   };
 
   const handleEdit = (item) => {
@@ -87,7 +100,7 @@ export default function AdminPublications() {
             <p><strong>Authors:</strong> {item.authors}</p>
             <p><strong>Journal:</strong> {item.journal}</p>
             <p>
-              <strong>Read More:</strong> 
+              <strong>Read More:</strong>
               <a href={item.link} target="_blank" rel="noopener noreferrer"> {item.link} </a>
             </p>
             <div className="action-buttons">
